fix: computing
